Fix region dimension inconsistency
This just got in the way of my debugging. You construct a region by providing width and length, yet when you put it in a string instead of width and height it outputs x/y screen position of lower right corner. I propose we display width and height instead when printing a region.